refactored calculate function on another file
