+++ homeworks/emiliano-acerbi/19-react-api/src/utils/calculate.js
@@ -1,0 +1,9 @@
+export const calculate = (a, b, sign) => {
+  const result = {
+    '+': (a, b) => a + b,
+    '-': (a, b) => a - b,
+    X: (a, b) => a * b,
+    '/': (a, b) => a / b,
+  };
+  return result[sign](a, b);
+};
